refactor(nav): collapse types and split eco-router tests
Export types directly from client/types, split eco-router tests into focused
files with a shared harness, and document the createRouter singleton contract.
